pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinelops"
version = "0.1.0"
description = "Explainable AI system for enterprise incident triage"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "networkx>=3.3",
    "sentence-transformers>=3.0.0",
    "pgvector>=0.3.0",
    "numpy>=1.26.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.35.0",
    "pandas>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "matplotlib>=3.9.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["sentinelops*"]
exclude = ["sentinelops.sentinelops*"]

[tool.setuptools.package-data]
sentinelops = ["simulation/runbooks/*.md"]
